+++ scripts-nino/test-magrpo-tldr-config.yaml
@@ -2,6 +2,9 @@
 
 # Model settings
 model_name: "Qwen/Qwen2.5-0.5B"
+
+lora_r: 32
+lora_alpha: 64
 target_modules: 
   - "q_proj"
   - "k_proj"
